fix: correct thousand separators and deduplicate parseUnixLocale
- French/Russian/Swedish/Finnish/Norwegian use non-breaking space (U+00A0)
  as thousand separator, not period
- Extract parseUnixLocale to shared regional_unix.go (!windows build tag)
  eliminating 3x duplication across linux/darwin/other
- Remove dead unsafe import from regional_windows.go
- Fix strings.Replace -> strings.ReplaceAll inconsistency in darwin
