+++ pkg/sqlcmd/regional_other.go
@@ -7,7 +7,6 @@ package sqlcmd
 
 import (
 	"os"
-	"strings"
 
 	"golang.org/x/text/language"
 )
@@ -26,39 +25,4 @@ func detectUserLocale() language.Tag {
 		}
 	}
 	return language.English
-}
-
-// parseUnixLocale converts a Unix locale string to a language.Tag
-// Examples: "en_US.UTF-8", "de_DE", "fr_FR.utf8", "C", "POSIX"
-func parseUnixLocale(locale string) language.Tag {
-	// Handle special cases
-	if locale == "C" || locale == "POSIX" || locale == "" {
-		return language.English
-	}
-
-	// Remove encoding suffix (e.g., ".UTF-8")
-	if idx := strings.Index(locale, "."); idx != -1 {
-		locale = locale[:idx]
-	}
-
-	// Remove modifier (e.g., "@euro")
-	if idx := strings.Index(locale, "@"); idx != -1 {
-		locale = locale[:idx]
-	}
-
-	// Convert underscore to hyphen for BCP 47 format
-	locale = strings.Replace(locale, "_", "-", -1)
-
-	if tag, err := language.Parse(locale); err == nil {
-		return tag
-	}
-
-	// Try with just the language part
-	if idx := strings.Index(locale, "-"); idx != -1 {
-		if tag, err := language.Parse(locale[:idx]); err == nil {
-			return tag
-		}
-	}
-
-	return language.Und
 }
